expenses/app/requirements.txt
fastapi
sqlalchemy
pydantic
python-jose
psycopg2-binary
kafka-python
